adapter refinements
added dashboard configuration tab for switching among the adapters
dynamically; changed adapter interface functions to camel casing and
updated some of the samples to use that syntax.  When running off local
filesystem and indexeddb is unavailable, the memory adapter is made the
default adapter and will works with dashboard without errors (even if
local storage is unavailable).